first pass at restructure
